@@ -10,10 +10,11 @@
 		"@fortawesome/free-solid-svg-icons": "^5.15.4",
 		"@fortawesome/react-fontawesome": "^0.1.16",
 		"@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
+		"@react-pdf/renderer": "3.3.5",
 		"@svgr/webpack": "^5.5.0",
 		"@types/node": "^16.11.21",
-		"@types/react": "^17.0.38",
-		"@types/react-dom": "^17.0.11",
+		"@types/react": "^18.2.61",
+		"@types/react-dom": "^18.2.19",
 		"babel-loader": "^8.2.3",
 		"babel-plugin-named-asset-import": "^0.3.8",
 		"babel-preset-react-app": "^10.0.1",
@@ -39,10 +40,10 @@
 		"postcss-normalize": "^10.0.1",
 		"postcss-preset-env": "^7.0.1",
 		"prompts": "^2.4.2",
-		"react": "^17.0.2",
+		"react": "^18.2.0",
 		"react-app-polyfill": "^3.0.0",
 		"react-dev-utils": "^12.0.0",
-		"react-dom": "^17.0.2",
+		"react-dom": "^18.2.0",
 		"react-refresh": "^0.11.0",
 		"react-tsparticles": "^1.39.1",
 		"resolve": "^1.20.0",
