@@ -1,17 +1,18 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import {library} from '@fortawesome/fontawesome-svg-core'
 import {fas} from '@fortawesome/free-solid-svg-icons'
 import {fab} from '@fortawesome/free-brands-svg-icons'
+import {createRoot} from 'react-dom/client'
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root')!)
+
+root.render(
 	<React.StrictMode>
 		<App />
-	</React.StrictMode>,
-	document.getElementById('root')
+	</React.StrictMode>
 )
 
 // If you want to start measuring performance in your app, pass a function
